+++ src/Movement/DDA.cpp
@@ -155,7 +155,7 @@ void DDA::DebugPrintVector(const char *name, const float *vec, size_t len) const
 	debugPrintf("%s=", name);
 	for (size_t i = 0; i < len; ++i)
 	{
-		debugPrintf("%c%f", ((i == 0) ? '[' : ' '), vec[i]);
+		debugPrintf("%c%f", ((i == 0) ? '[' : ' '), (double)vec[i]);
 	}
 	debugPrintf("]");
 }
@@ -175,12 +175,12 @@ void DDA::DebugPrint() const
 		DebugPrintVector(" end", endCoordinates, numAxes);
 	}
 
-	debugPrintf(" d=%f", totalDistance);
+	debugPrintf(" d=%f", (double)totalDistance);
 	DebugPrintVector(" vec", directionVector, 5);
 	debugPrintf("\na=%f reqv=%f topv=%f startv=%f endv=%f\n"
 				"daccel=%f ddecel=%f cks=%u\n",
-				acceleration, requestedSpeed, topSpeed, startSpeed, endSpeed,
-				accelDistance, decelDistance, clocksNeeded);
+				(double)acceleration, (double)requestedSpeed, (double)topSpeed, (double)startSpeed, (double)endSpeed,
+				(double)accelDistance, (double)decelDistance, clocksNeeded);
 	for (size_t axis = 0; axis < numAxes; ++axis)
 	{
 		if (pddm[axis] != nullptr)
@@ -1028,12 +1028,12 @@ void DDA::Prepare(uint8_t simMode)
 					{
 						// The distance to reversal is the solution to a quadratic equation. One root corresponds to the carriages being below the bed,
 						// the other root corresponds to the carriages being above the bed.
-						const float drev = ((directionVector[Z_AXIS] * sqrt(a2b2D2 - fsquare(A * directionVector[Y_AXIS] - B * directionVector[X_AXIS])))
+						const float drev = ((directionVector[Z_AXIS] * sqrtf(a2b2D2 - fsquare(A * directionVector[Y_AXIS] - B * directionVector[X_AXIS])))
 											- aAplusbB)/a2plusb2;
 						if (drev > 0.0 && drev < totalDistance)		// if the reversal point is within range
 						{
 							// Calculate how many steps we need to move up before reversing
-							const float hrev = directionVector[Z_AXIS] * drev + sqrt(dSquaredMinusAsquaredMinusBsquared - 2 * drev * aAplusbB - a2plusb2 * fsquare(drev));
+							const float hrev = directionVector[Z_AXIS] * drev + sqrtf(dSquaredMinusAsquaredMinusBsquared - 2 * drev * aAplusbB - a2plusb2 * fsquare(drev));
 							const int32_t numStepsUp = (int32_t)((hrev - h0MinusZ0) * stepsPerMm);
 
 							// We may be almost at the peak height already, in which case we don't really have a reversal.
